Fix custom element state setting for new CustomElement()
As discussed in #1297, the spec previously did not correctly set custom
element state to "custom" for the case of new CustomElement(). The
correct place to do this is in the HTMLElement constructor (invoked by
the super() call), since that is the last place user agent code is
invoked, and so our last chance to change the state. In practice, this
affects the behavior of the :defined pseudo-class.

Fixes #1297. Also fixes a typo in the same algorithms, where the
variable names "element" and "interface" were accidentally both used.